# api-quiz/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/quiz/exam/users/join": {
      "post": {
        "operationId": "joinExam",
        "summary": "",
        "description": "Join exam",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamUserInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam.users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/quiz/exam/users/leave": {
      "post": {
        "operationId": "leaveExam",
        "summary": "",
        "description": "Leave exam",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamUserInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam.users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/quiz/exam/users/add": {
      "post": {
        "operationId": "addMember",
        "summary": "",
        "description": "Add member",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamUserAdminInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam.users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/quiz/exam/users/remove": {
      "post": {
        "operationId": "removeUser",
        "summary": "",
        "description": "Remove user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamUserAdminInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam.users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/quiz/admin/exam/users/add": {
      "post": {
        "operationId": "addMember",
        "summary": "",
        "description": "Add member",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamUserAdminInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin.exam.users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/quiz/admin/exam/users/remove": {
      "post": {
        "operationId": "removeUser",
        "summary": "",
        "description": "Remove user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamUserAdminInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin.exam.users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/api/quiz/exam/questions/add": {
      "post": {
        "operationId": "addMember",
        "summary": "",
        "description": "Add member",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamQuestionAdminInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam.questions"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/quiz/exam/questions/remove": {
      "post": {
        "operationId": "removeUser",
        "summary": "",
        "description": "Remove user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamQuestionAdminInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "exam.questions"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/quiz/admin/exam/questions/add": {
      "post": {
        "operationId": "addMember",
        "summary": "",
        "description": "Add member",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamQuestionAdminInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin.exam.questions"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/quiz/admin/exam/questions/remove": {
      "post": {
        "operationId": "removeQuestion",
        "summary": "",
        "description": "Remove Question",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamQuestionAdminInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin.exam.questions"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Quiz API",
    "description": "Swagger specification for api-quiz | [swagger.json](swagger.json)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "message": {
            "type": "string",
            "example": "Error message",
            "default": "Internal Server Error"
          },
          "name": {
            "type": "string",
            "example": "Error"
          }
        }
      },
...
      "CreateExamUserInput": {
        "type": "object",
        "properties": {
          "examId": {
            "type": "string",
            "description": "Exam Id",
            "example": "6399f5469f80115a1fb3031c"
          }
        },
        "required": [
          "examId"
        ]
      },
      "CreateExamUserAdminInput": {
        "type": "object",
        "properties": {
          "examId": {
            "type": "string",
            "description": "Exam Id",
            "example": "6399f5469f80115a1fb3031c"
          },
          "userId": {
            "type": "string",
            "description": "User Id",
            "example": "6399f5469f80115a1fb3031c"
          }
        },
        "required": [
          "examId",
          "userId"
        ]
      },
...
      "CreateExamQuestionAdminInput": {
        "type": "object",
        "properties": {
          "examId": {
            "type": "string",
            "description": "Exam Id",
            "example": "6399f5469f80115a1fb3031c"
          },
          "questionId": {
            "type": "string",
            "description": "Question Id",
            "example": "6399f5469f80115a1fb3031c"
          }
        },
        "required": [
          "examId",
          "questionId"
        ]
      }
    }
  }
}